feat(suite-common): Refactor useTimer to allow longer ticks
