+++ packages/react-utils/src/hooks/timer/Timer.ts
@@ -1,0 +1,11 @@
+export interface Timer {
+    timeSpent: {
+        seconds: number;
+    };
+    resetCount: number;
+    isStopped: boolean;
+    isLoading: boolean;
+    stop: () => void;
+    reset: () => void;
+    loading: () => void;
+}
