+++ packages/react-utils/src/index.ts
@@ -4,10 +4,9 @@ export { useDidUpdate } from './hooks/useDidUpdate';
 export { useKeyPress } from './hooks/useKeyPress';
 export { useOnce } from './hooks/useOnce';
 export { useOnClickOutside } from './hooks/useOnClickOutside';
-export { useTimer } from './hooks/useTimer';
 export { useWindowFocus } from './hooks/useWindowFocus';
 export { useClickCooldown } from './hooks/useClickCooldown';
 export { useAsyncClickHandler } from './hooks/useAsyncClickHandler';
 export { usePreviousDefined } from './hooks/usePreviousDefined';
 export { useTextareaCursorPosition } from './hooks/useTextareaCursorPosition';
-export type { Timer } from './hooks/useTimer';
+export * from './hooks/timer';
